Revert "Remove IT test for session because it's tested on JS side"
This reverts commit c8a23c3f2216fb9d8b8c90ae511944551fdf4d0b.

Restoring these tests because we need to preserve the old session.html
behavior to maintain backwards compatibility with old uaa-singulars
that don't support session management.

With the introduction of session management, the string being passed
from the RP iframe to OP iframe will change in a non-backwards
compatible way so UAA needs to provide iframe content for both formats.

[#153568058] https://www.pivotaltracker.com/story/show/153568058

Signed-off-by: Mikhail Vyshegorodtsev <medvedzver@inbox.ru>